+++ lib/view_component/render_component_helper.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module ViewComponent
+  module RenderComponentHelper # :nodoc:
+    def render_component(component, &block)
+      component.render_in(self, &block)
+    end
+  end
+end