+++ test/config/environments/test.rb
@@ -21,6 +21,8 @@ Dummy::Application.configure do
 
   config.view_component.show_previews = true
 
+  config.view_component.render_monkey_patch_enabled = true
+
   config.view_component.test_controller = "IntegrationExamplesController"
 
   # Tell Action Mailer not to deliver emails to the real world.